Read directly from window and provide more plumbable errors
This change makes `diff` compare the formatter output to the body of
the acme window, rather than the file on disk with the same name. This
way there are no surprises if what is on disk does not match what is in the
acme window (for example, if the file is being modified by another program).

In addition, error output from the formatter is logged to a +Errors window
sharing the same prefix as the file being formatted. If the formatter prints
any line numbers or addresses, they can then be plumbed to jump to the
location in the file.